fix linting errors: 'safe' changes.
This change replaces timer.Tick with timer.NewTicker (and defers
a call to timer.Stop to avoid theoretically leaking goroutines.
The current code seems to be fine as is, because the loops are
long running and can not be exited at the moment, but I think it is
still good habit to write well-behaving code.

Additionally a small buffer is added to the os.Signal channel
(according to the docs, those channels must be buffered because
the signal handler can not block and can not do a synchronized
send).

Neither of those changes should change the behavior in any way.