Use tap instead of tun
The yggdrasil-network/water patch to make water work on bsd does some
confusing things. One thing is that unless you set `Name`, it will
create a tap device. Fix this by requiring the user set `tun.dev` in
their config file to a name like `tun0` or `tun1` (the naming required
by freebsd). We also need to add the `route` command back now that we
are using a tun device.