SSH: don't hold read locks longer than necessary
Previously we were holding hostmap and lighthouse read locks for the
duration of the commands `list-hostname` and `list-lighthouse-addrmap`
(including the network write time). For busy lighthouses this could
block host updates for a significant amount of time.

Now we make a copy of the data we're going to serialize and then
release the lock before marshaling it and sending it back down to the
ssh client.