Initial implementation of experimental plugin architecture
This commit adds a new Plugin interface type and the associated plumbing
to enable modularised features to be experimented with in the form of go
plugins. Any plugin that implements the interface can be loaded at startup
and register to send/receive message packets with a configurable subType.

All packets sent by a plugin are prefixed with the shell of an IPv4 header
to allow for nebula firewall rules to appropriately allow/deny payloads.
Received packets are processed by the firewall and then passed to the plugin
via its Receive method with the IPv4 header bytes included.

Plugins can accept configuration options in the form of a map, which is
included in the `experimental.plugins` section of the main config file.