plugins: Reference implementation
This commit adds a new plugin called 'echo'. When the plugin receives a
packet, it will write the payload as a string to the host nebula log.
When appropriate config is provided, the plugin will send the requested
payload to the configured nebula IP address at regular intervals.

The plugin must be built using the same go version and environment as
nebula core via:
`go build -buildmode=plugin -o nebula_plugin_echo.so ./plugin/echo`

Sample config for sender (omit the `config` sub section on the receiver):

```
experimental:
  plugins:
    - name: echo
      path: nebula_plugin_echo.so
      id: 1
      config:
        dest: 192.168.100.1
        payload: hello lighthouse
        period: 3s
```