Refuse to load plugins when running as root
The ability to load arbitrary plugins makes is easy to pwn a host that
has nebula running as root/Administrator. This commit adds a safety
net that refuses to load plugins unless nebula is running as a user
account that is not root/Admin, or the exact risk acceptance statement
is added to the config.

This can't protect against a scenario where nebula is running as a
non-root user with sudo privileges, so mabe it's not even valuable?