Add ability to encrypt CA private key at rest
Fixes #8. This commit causes `nebula-cert ca` to prompt for a passphrase
by default. If a passphrase is entered, the private key will be
encrypted using AES-256 and the Argon2id key derivation function. A new
private key banner has been added for encrypted private keys. A flag
`-no-encryption` can also be passed to skip this prompt (e.g. for
scripting.)

Calls to `nebula-cert sign` which point to an encrypted CA private key
will now prompt for a passphrase. Unencrypted CA private keys will not
be prompted for a passphrase.