even better fix
Instead of locking, just set the hostId before we add the hostinfo to
the map. This way it is static and never changes, and we don't need to
lock around reading it.

Do the same thing for remoteIndex because we also know this initially
and we don't need to lock around the map again to re-add it.