we don't need handshakeManager for this
For IX, we know everything we need and we are going to add this host to
the real hostmap at the end of the function. So there is no need to add
to the pending hostmap which is just going to require the inbound timer
tick to come back and reap it later.