Only set serveDns if the host is also configured to be a lighthouse.
Nebula attempts an Add() call against the global dnsR variable if serveDns is true.
Unfortunately, the global dnsR is only initialized in main.go if the dns config is set to true AND the system is configured to be a lighthouse.
If the host is not configured to be a lighthouse but is also configured to serve DNS, the Add call will panic against the nil dnsR.