Fix UDP listener on IPv4-only Linux
On some systems, IPv6 is disabled (for example, CIS benchmark recommends to disable it when not used), but currently all UDP connections are using AF_INET6 sockets.
When we are binding AF_INET6 socket to an address like ::ffff:1.2.3.4 (IPv4 addresses are parsed by net.Listen this way), we can't send or receive IPv6 packets anyway, so this will not break any scenarios.

Fixes #467