Add README and restart burst limit to service file
Add a 30 second delay before restarting the service after a failure and
prevents additional restarts after 5 attempts.

Add some basic usage/installation instructions for the Systemd service
file to the README.md within the example service file directory.

Change `After=` to nss-lookup for when hostnames are in the nebula config file and change `Wants=` to network-online