Load expired CAs into the pool, but log about it
It's arguably better for the expired CAs to be loaded into the pool
so that connections from hosts using that CA will log that the root
is expired, rather than being missing (which would be a behaviour
change caused only by a daemon restart/reload).

This iteration will load all otherwise-valid certs into the CA pool
and propagate an ErrExpired to be handled by the caller. If the pool
contains only expired certs or any other invalid cert is present in
the PEM, we return a different error and fail to start/reload.