refactor: break apart config reading from Config.Load
This `Config` struct was responsible for reading files from disk
and parse as YAML. This prevented to ReloadConfig when no files were
passed, i.e., only the marshalled YAML config was used.

This commit breaks the whole file loading and lexicographic sorting in a
separate exported function in the package, called
config.ReadConfigFiles(path string).

This commit adds a few tests to cover sorting and mixed files loading
checks.