add listen.send_recv_error config option
By default, Nebula replies to packets it has no tunnel for with a `recv_error` packet. This packet helps speed up reconnection
in the case that Nebula on either side did not shut down cleanly. This response can be abused as a way to discover if Nebula is running
on a host though. This option lets you configure if you want to send "recv_error" packets always, never, or only to private network remotes.
valid values: true, false, private

This setting is reloadable with SIGHUP.