add punchy.respond_delay config option
Currently, the delay before starting the punchy.respond attempt is
hardcoded to 5 seconds. If you know your host is behind a problematic
NAT, you may want to configure this lower. This change lets you do this.

I also removed a comment from "punchy.delay" that was incorrect (the
punchy.delay option is not dependany on punchy.respond being set)