Immediately forward packets from self to self on FreeBSD
Fixes #493. This is borrowing the fix from #501 which was originally
scoped to use this behavior for all OSes. During review, it was scoped
down to just macOS which was known to be affected. Lacking support for
FreeBSD at the time was an oversight.

There may be other OSes which are affected by this bug, but we will wait
until we get reports to patch them.